requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
# CPU-only runtime requirements for CI/image builds.
# Use CNN/requirements.txt for CUDA training dependencies.
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
fastapi==0.108.0
uvicorn[standard]==0.24.0.post1
httpx==0.27.2
python-multipart==0.0.22
pyyaml==6.0.3
pillow==12.1.0
pillow-heif==0.20.0
scikit-learn==1.8.0
matplotlib==3.10.8
# NOTE: onnxruntime 1.17.x wheels are not compatible with NumPy 2.x.
# Pin NumPy <2 to avoid import/runtime errors (CI + local).
numpy==1.26.4
tqdm==4.66.1
pytest==8.2.0
timm==1.0.3
pandas==2.2.3
onnxruntime==1.17.3
